Setup Duke package
