+++ src/main/java/Main.java
@@ -1,0 +1,8 @@
+import Duke.Duke;
+
+public class Main {
+    public static void main(String[] args) {
+        Duke duke = new Duke();
+        duke.run();
+    }
+}
